requirement.txt
langchain
langchain-community
pypdf
pymupdf
langchain-text-splitters
langchain-google-genai
python-dotenv
faiss-cpu